Major update
Added back-end functionality. Added configurations fpr uwsgi and nginx. Started developing front-end part

+++ messageExchangeACS/department/views.py
@@ -1,4 +1,3 @@
-from django.core.exceptions import ObjectDoesNotExist
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -6,25 +5,23 @@ from rest_framework import status
 
 from .serializers import DepartmentSerializer
 from .models import Department
-from institute.models import Institute
 
 
 class InstituteView(APIView):
     permission_classes = (IsAuthenticated,)
 
-    def get_object(self, department_id):
-        try:
-            return Department.objects.get(pk=department_id)
-        except ObjectDoesNotExist:
-            return None
+    def get(self, request, department_id=None, format=None):
+        if department_id:
+            department = Department.get_by_id(department_id)
+            if not department:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+            serializer = DepartmentSerializer(department)
+            return Response(serializer.data)
+        departments = Department.objects.all()
+        response = DepartmentSerializer(departments, many=True)
+        return Response(response.data)
 
-    def get(self, request, department_id, format=None):
-        department = self.get_object(department_id)
-        serializer = DepartmentSerializer(department)
-        return Response(serializer.data)
-
-    def post(self, request, institute_id=None, format=None):
-        # institute = Institute.get_by_id(institute_id)
+    def post(self, request, format=None):
         serializer = DepartmentSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -32,7 +29,7 @@ class InstituteView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, department_id, format=None):
-        department = self.get_object(department_id)
+        department = Department.get_by_id(department_id)
         serializer = DepartmentSerializer(department, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -40,6 +37,8 @@ class InstituteView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, department_id, format=None):
-        department = self.get_object(department_id)
+        department = Department.get_by_id(department_id)
+        if not department:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         department.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
