+++ messageExchangeACS/institute/views.py
@@ -1,4 +1,3 @@
-from django.core.exceptions import ObjectDoesNotExist
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -11,15 +10,11 @@ from .models import Institute
 class InstituteView(APIView):
     permission_classes = (IsAuthenticated,)
 
-    def get_object(self, institute_id):
-        try:
-            return Institute.objects.get(pk=institute_id)
-        except ObjectDoesNotExist:
-            return None
-
     def get(self, request, institute_id=None, format=None):
         if institute_id:
-            institute = self.get_object(institute_id)
+            institute = Institute.get_by_id(institute_id)
+            if not institute:
+                return Response(status=status.HTTP_404_NOT_FOUND)
             serializer = InstituteSerializer(institute)
             return Response(serializer.data)
         institutes = Institute.objects.all()
@@ -27,7 +22,6 @@ class InstituteView(APIView):
         return Response(response.data)
 
     def post(self, request, format=None):
-        print(request.data)
         serializer = InstituteSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -35,7 +29,9 @@ class InstituteView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, institute_id, format=None):
-        institute = self.get_object(institute_id)
+        institute = Institute.get_by_id(institute_id)
+        if not institute:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = InstituteSerializer(institute, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -43,6 +39,8 @@ class InstituteView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, institute_id, format=None):
-        institute = self.get_object(institute_id)
+        institute = Institute.get_by_id(institute_id)
+        if not institute:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         institute.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
